로그아웃 로직 추가, 로그인 버튼 눌렀을 시 로그인 button 누를 시 Modal로 로그인 버튼 표현
