+++ sac/src/components/Header/Header.js
@@ -3,10 +3,14 @@ import React, { useState, useEffect } from 'react';
 import searchIcon from '../../assets/image/search.png';
 import '../../styles/Header.css';
 import UserInfo from '../User/UserInfo';
+import Modal from 'react-modal'; // 모달 라이브러리 import
+import GoogleLogInButton from '../../components/Login/GoogleLogInButton';
+import customModalStyles from '../../styles/Modal';
 
 function Header() {
     const [searchText, setSearchText] = useState('');
     const [isLoggedIn, setIsLoggedIn] = useState(false);
+    const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
 
     useEffect(() => {
         // localStorage에서 loginSession 확인
@@ -16,6 +20,14 @@ function Header() {
 
     const handleSearchChange = (event) => {
         setSearchText(event.target.value);
+    };
+
+    const openLoginModal = () => {
+        setIsLoginModalOpen(true);
+    };
+
+    const closeLoginModal = () => {
+        setIsLoginModalOpen(false);
     };
 
     return (
@@ -44,10 +56,19 @@ function Header() {
                     {isLoggedIn ? (
                         <UserInfo />
                     ) : (
-                        <button className='login-button'>로그인</button>
+                        <button className='login-button' onClick={openLoginModal}>로그인</button>
                     )}
                 </div>
             </div>
+            {/* 모달 영역 */}
+            <Modal
+                isOpen={isLoginModalOpen}
+                onRequestClose={closeLoginModal}
+                style={customModalStyles}
+                // 모달 스타일 등 추가 설정 가능
+            >
+                <GoogleLogInButton />
+            </Modal>
         </div>
     );
 }
